back/pyproject.toml
[tool.poetry]
name = "broadcast"
version = "0.1.0"
description = ""
authors = ["Robert Rakhmatullin <robertbg4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-phonenumber-field = "^5.0.0"
Django = "^3.1.5"
phonenumbers = "^8.12.16"
djangorestframework = "^3.12.2"
requests = "^2.25.1"
django-cors-headers = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
